wallet page topup added

+++ Modules/PromoRules/database/migrations/2025_09_10_074339_alter_wallet_lots_table.php
@@ -14,7 +14,9 @@ return new class extends Migration
             $table->decimal('base_value', 12, 2)->after('amount');
             $table->decimal('bonus_value', 12, 2)->default(0)->after('base_value');
 //            $table->foreignId('gift_card_id')->nullable()->constrained('st_gift_cards')->onDelete('set null')->after('bonus_value');
-            $table->foreignId('promo_rule_id')->nullable()->constrained('promo_rules')->onDelete('set null')->after('gift_card_id');
+            if (!Schema::hasColumn('wallet_lots', 'promo_rule_id')) {
+                $table->foreignId('promo_rule_id')->nullable()->constrained('promo_rules')->onDelete('set null')->after('gift_card_id');
+            }
         });
     }
